// src/app/components/web3/types/poi.ts
// Proof of Identity

export type CreatePoi = {
  email: string;
  fullName: string;
  username: string;
  phone: string;
}

export type Poi = {
  email: string;
  fullName: string;
  username: string;
  phone: string;
  country: string;
  gender: string;
  dateOfBirth: string;
  imageLink: string;
  fbLink: string;
  igLink: string;
  youtubeLink: string;
  yTWelcomeLink: string;
  tikTokLink: string;
  wspLink: string;
  bio: string;
}